Add DownloadService for portal downloads
DownloadManager cannot be used by the captive portal application because
it does not allow specifying a network. Because of this a download using
DownloadManager would break connectivity on the device, as a network
without internet access would need to be set as default.

Add a dedicated DownloadService to the application so it can do its own
downloads in a foreground service.

Close icon source:
https://material.io/resources/icons/?icon=close&style=baseline

Download icon source:
https://material.io/resources/icons/?icon=cloud_download&style=baseline

Bug: 131401028
Test: atest CaptivePortalLoginTests
Change-Id: Ibb449023f2559a27113d105f5d38f1fc2148edc9