Use DownloadService in CaptivePortalLoginActivity
This allows the activity to download files when WebView sees download
attempts.
This change also allows applications to be downloaded from the portal,
although users will need to whitelist the app for installing APKs from
settings (they are prompted to do so when attempting downloads).
Applications may be offered on portals such as airline wifi, in which
case installing the app from the portal can be useful.

Bug: 131401028
Test: manual + tests in I92d13f6dd281f198da86d7a9f37bff7571a318d7
Change-Id: I8b8b27573ab2ed3ee5c97e6420655378384bc7e0