Mock the MimeTypeMap for DownloadServiceTest
Some devices have an application that handles all intents for the
application/octet-stream mimetype, which breaks testTapDoneNotification
because an app chooser dialog is shown instead of opening the
OpenTextFileActivity directly.

To avoid such problems, mock MimeTypeMap to register a custom mimetype
that is highly unlikely to be used by any application. DownloadService
will then detect that mimetype from the test file extension, and only
OpenTextFileActivity will handle the intent.

Also remove the RequestDismissKeyguardActivity in DownloadServiceTest,
which was a duplicate (minus the dismiss behavior) of the one in its own
file in the same package.

Bug: 162296431
Test: atest CaptivePortalLoginTests:DownloadServiceTest
Change-Id: I439c639fe0dced53ad506af03facdf3f8ae07199