Scroll to find hidden DownloadService notification
If the test is run on a device with a lot of notifications and/or
limited vertical space, it would fail as the DownloadService
notification was not in the screen.

Optimistically try to find notifications with a short timeout, then
scroll in the notification shade to find it after the short timeout.
Close the notification shade and retry until the test timeout while it
is not found.

Test: atest CaptivePortalTests:DownloadServiceTest
Bug: 162296431
Change-Id: I8d6f1dbe97c5e8561bd7605298a63e79cf5b0c16