Add a version check before calling isAlwaysOnVpnLockdownEnabled()
There is a permission check change in DevicePolicyManagerService
in Android R which allows the caller who has the
PERMISSION_MAINLINE_NETWORK_STACK pass the permission check.
CaptivePortalLogin is one of the callers of that API and
CaptivePortalLogin is also one of the mainline modules, which
means the changes in CaptivePortalLogin may be also updated to
Android Q device. But Android Q doesn't allow CaptivePortalLogin
pass the permission check, and it results in CaptivePortalLogin
crash.

Bug: 168567727
Test: 1. atest CaptivePortalLoginActivityTest
      2. Check the issue could be reproduced by installing the
         latest CaptivePortalLogin.apk into Android Q device.
      3. Check the issue could not be reproduced by installing the
         CaptivePortalLogin.apk with the solution into Android Q
         device.
Change-Id: I49f40a228a39eca51f2eee7bbe2cd64100d6ea46