Use FLAG_IMMUTABLE in download done notification
Without the flag, the app will crash when trying to show the
notification once its target SDK is updated to 31+.

Also update target SDK in tests, and add comments to make sure it is
kept in sync with the actual module.

Fix DownloadServiceTest to verify the change, as it verifies that "1%"
is shown in the notification, which is not necessarily true depending on
the notification UI, especially in S. Remove this assertion, which can
be wrong.

Bug: 191084568
Test: atest DownloadServiceTest
Change-Id: If97ac78d0b89f1cabec6bf5d967f71135430751f